src/auth_demo/backend/requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
ordered-set==4.1.0
PyJWT==2.10.1
python-dotenv==1.2.1
Werkzeug==3.1.3
